+++ src/Ydb.Sdk/test/Ydb.Sdk.Ado.Tests/YdbParameterTests.cs
@@ -222,7 +222,7 @@ public class YdbParameterTests : TestBase
     public async Task YdbParameter_WhenYdbDbTypeSetAndValueIsNotNull_ReturnsValue()
     {
         await using var ydbConnection = await CreateOpenConnectionAsync();
-        var tableName = $"NouNull_YdbDbType_{Random.Shared.Next()}";
+        var tableName = $"NonNull_YdbDbType_{Random.Shared.Next()}";
         await new YdbCommand(ydbConnection)
         {
             CommandText = $"""
@@ -248,6 +248,10 @@ public class YdbParameterTests : TestBase
                                IntervalColumn Interval NOT NULL,
                                JsonColumn Json NOT NULL,
                                JsonDocumentColumn JsonDocument NOT NULL,
+                               Date32Column Date32 NOT NULL,
+                               Datetime64Column DateTime64 NOT NULL,
+                               Timestamp64Column Timestamp64 NOT NULL,
+                               Interval64Column Interval64 NOT NULL,
                                PRIMARY KEY (Int32Column)
                            );
                            """
@@ -260,12 +264,14 @@ public class YdbParameterTests : TestBase
                                Int32Column, BoolColumn, Int64Column, Int16Column, Int8Column, FloatColumn, DoubleColumn, 
                                DefaultDecimalColumn, CustomDecimalColumn, Uint8Column, Uint16Column, Uint32Column, 
                                Uint64Column, TextColumn, BytesColumn, DateColumn, DatetimeColumn, TimestampColumn,
-                               IntervalColumn, JsonColumn, JsonDocumentColumn
+                               IntervalColumn, JsonColumn, JsonDocumentColumn, Date32Column, Datetime64Column,
+                               Timestamp64Column,  Interval64Column
                            ) VALUES (
                                @Int32Column, @BoolColumn, @Int64Column, @Int16Column, @Int8Column, @FloatColumn, 
                                @DoubleColumn, @DefaultDecimalColumn, @CustomDecimalColumn, @Uint8Column, @Uint16Column, 
                                @Uint32Column, @Uint64Column, @TextColumn, @BytesColumn, @DateColumn, @DatetimeColumn, 
-                               @TimestampColumn, @IntervalColumn, @JsonColumn, @JsonDocumentColumn
+                               @TimestampColumn, @IntervalColumn, @JsonColumn, @JsonDocumentColumn, @Date32Column,
+                               @Datetime64Column,  @Timestamp64Column, @Interval64Column
                            );
                            """,
             Parameters =
@@ -290,7 +296,12 @@ public class YdbParameterTests : TestBase
                 new YdbParameter("TimestampColumn", YdbDbType.Timestamp, DateTime.UnixEpoch),
                 new YdbParameter("IntervalColumn", YdbDbType.Interval, TimeSpan.Zero),
                 new YdbParameter("JsonColumn", YdbDbType.Json, "{}"),
-                new YdbParameter("JsonDocumentColumn", YdbDbType.JsonDocument, "{}")
+                new YdbParameter("JsonDocumentColumn", YdbDbType.JsonDocument, "{}"),
+                new YdbParameter("Date32Column", YdbDbType.Date32, DateTime.MinValue),
+                new YdbParameter("Datetime64Column", YdbDbType.Datetime64, DateTime.MinValue),
+                new YdbParameter("Timestamp64Column", YdbDbType.Timestamp64, DateTime.MinValue),
+                new YdbParameter("Interval64Column", YdbDbType.Interval64,
+                    TimeSpan.FromMilliseconds(TimeSpan.MinValue.Milliseconds))
             }
         }.ExecuteNonQueryAsync();
 
@@ -301,7 +312,8 @@ public class YdbParameterTests : TestBase
                                Int32Column, BoolColumn, Int64Column, Int16Column, Int8Column, FloatColumn, DoubleColumn, 
                                DefaultDecimalColumn, CustomDecimalColumn, Uint8Column, Uint16Column, Uint32Column, 
                                Uint64Column, TextColumn, BytesColumn, DateColumn, DatetimeColumn, TimestampColumn,
-                               IntervalColumn, JsonColumn, JsonDocumentColumn 
+                               IntervalColumn, JsonColumn, JsonDocumentColumn,  Date32Column, Datetime64Column,  
+                               Timestamp64Column, Interval64Column
                            FROM {tableName};
                            """
         }.ExecuteReaderAsync();
@@ -328,14 +340,13 @@ public class YdbParameterTests : TestBase
         Assert.Equal(TimeSpan.Zero, ydbDataReader.GetInterval(18));
         Assert.Equal("{}", ydbDataReader.GetJson(19));
         Assert.Equal("{}", ydbDataReader.GetJsonDocument(20));
+        Assert.Equal(DateTime.MinValue, ydbDataReader.GetDateTime(21));
+        Assert.Equal(DateTime.MinValue, ydbDataReader.GetDateTime(22));
+        Assert.Equal(DateTime.MinValue, ydbDataReader.GetDateTime(23));
+        Assert.Equal(TimeSpan.FromMilliseconds(TimeSpan.MinValue.Milliseconds), ydbDataReader.GetInterval(24));
         Assert.False(ydbDataReader.Read());
         await ydbDataReader.CloseAsync();
 
         await new YdbCommand(ydbConnection) { CommandText = $"DROP TABLE {tableName}" }.ExecuteNonQueryAsync();
-    }
-
-    [Fact]
-    public async Task YdbDbType_WhenEnableNewDateTimeTypes_ReturnsExpectedTypes()
-    {
     }
 }